+++ src/config/config.ts
@@ -3,9 +3,8 @@ import WSL from 'is-wsl'
 import {arch, userInfo as osUserInfo, release, tmpdir, type} from 'node:os'
 import {join, sep} from 'node:path'
 import {URL, fileURLToPath} from 'node:url'
-import {format} from 'node:util'
 
-import {stdout} from '../cli-ux/stream'
+import {ux} from '../cli-ux'
 import {Command} from '../command'
 import {CLIError, error, exit, warn} from '../errors'
 import {getHelpFlagAdditions} from '../help/util'
@@ -499,7 +498,7 @@ export class Config implements IConfig {
           exit(code)
         },
         log(message?: any, ...args: any[]) {
-          stdout.write(format(message, ...args) + '\n')
+          ux.info(message, ...args)
         },
         warn(message: string) {
           warn(message)
@@ -528,7 +527,17 @@ export class Config implements IConfig {
         } catch (error: any) {
           final.failures.push({error: error as Error, plugin: p})
           debug(error)
-          if (!captureErrors && error.oclif?.exit !== undefined) throw error
+          // Do not throw the error if
+          // captureErrors is set to true
+          // error.oclif.exit is undefined or 0
+          // error.code is MODULE_NOT_FOUND
+          if (
+            !captureErrors &&
+            error.oclif?.exit !== undefined &&
+            error.oclif?.exit !== 0 &&
+            error.code !== 'MODULE_NOT_FOUND'
+          )
+            throw error
         }
 
         marker?.addDetails({