+++ src/config/ts-node.ts
@@ -1,11 +1,10 @@
-import {existsSync} from 'node:fs'
 import {join, relative as pathRelative, sep} from 'node:path'
 import * as TSNode from 'ts-node'
 
 import {memoizedWarn} from '../errors'
 import {Plugin, TSConfig} from '../interfaces'
 import {settings} from '../settings'
-import {readJsonSync} from '../util/fs'
+import {existsSync, readJsonSync} from '../util/fs'
 import {isProd} from '../util/util'
 import Cache from './cache'
 import {Debug} from './util'
@@ -24,7 +23,7 @@ function loadTSConfig(root: string): TSConfig | undefined {
     typescript = require('typescript')
   } catch {
     try {
-      typescript = require(join(root, 'node_modules', 'typescript'))
+      typescript = require(require.resolve('typescript', {paths: [root, __dirname]}))
     } catch {
       debug(`Could not find typescript dependency. Skipping ts-node registration for ${root}.`)
       memoizedWarn(
@@ -201,7 +200,7 @@ function determinePath(root: string, orig: string): string {
 export function tsPath(root: string, orig: string, plugin: Plugin): string
 export function tsPath(root: string, orig: string | undefined, plugin?: Plugin): string | undefined
 export function tsPath(root: string, orig: string | undefined, plugin?: Plugin): string | undefined {
-  const rootPlugin = Cache.getInstance().get('rootPlugin')
+  const rootPlugin = plugin?.options.isRoot ? plugin : Cache.getInstance().get('rootPlugin')
 
   if (!orig) return orig
   orig = orig.startsWith(root) ? orig : join(root, orig)
@@ -223,10 +222,16 @@ export function tsPath(root: string, orig: string | undefined, plugin?: Plugin):
       memoizedWarn(
         `${plugin?.name} is a linked ESM module and cannot be auto-transpiled. Existing compiled source will be used instead.`,
       )
+
+    if (plugin?.options.url)
+      memoizedWarn(
+        `${plugin?.name} is an ESM module installed from github and cannot be auto-transpiled. Existing compiled source will be used instead.`,
+      )
     return orig
   }
 
-  if (settings.tsnodeEnabled === undefined && isProduction && plugin?.type !== 'link') {
+  // Do not skip ts-node registration if the plugin is linked or installed from github
+  if (settings.tsnodeEnabled === undefined && isProduction && plugin?.type !== 'link' && !plugin?.options.url) {
     debug(`Skipping ts-node registration for ${root} because NODE_ENV is NOT "test" or "development"`)
     return orig
   }